ai-appointment-system/requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
flask-sqlalchemy==2.5.1
datetime
pytz
google-generativeai
flask-babel